refactor(products): Remove the `useProducts` hook and its dependencies
- Files related to the `useProducts` hook have been removed, including mutations, queries, types, and utilities, simplifying the code structure and eliminating unnecessary dependencies.
- This removal contributes to better code organization and improved project maintainability.
