--- app/store/hooks/data/useProducts/index.ts
@@ -1,14 +1,0 @@
-// Exportar el hook principal
-export { useProducts } from './useProducts';
-
-// Exportar tipos
-export type * from './types';
-
-// Exportar utilidades
-export * from './utils';
-
-// Exportar mutaciones
-export * from './mutations';
-
-// Exportar queries
-export * from './queries';
