--- app/store/hooks/data/useProducts/mutations/index.ts
@@ -1,1 +1,0 @@
-export { useProductMutations } from './useProductMutations';
