--- app/store/hooks/data/useProducts/queries/index.ts
@@ -1,1 +1,0 @@
-export { useProductQueries } from './useProductQueries';
