--- app/store/hooks/data/useProducts/types/index.ts
@@ -1,1 +1,0 @@
-export type * from './useProducts-types';
