--- app/store/hooks/data/useProducts/utils/index.ts
@@ -1,2 +1,0 @@
-export { useProductCacheUtils } from './productCacheUtils';
-export { useProductPagination } from './productPaginationUtils';
